+++ tests/test_tabula.py
@@ -1,5 +1,3 @@
-from teaser.project import Project
-prj = Project(False)
 
 
 class Test_tabula(object):
@@ -956,369 +954,8 @@ class Test_tabula(object):
             sum(wall.area for wall in
                 prj.buildings[-1].thermal_zones[-1].doors), 1) == 2.0
 
-    def test_tabula_uvalue_standard_sfh(self):
-        prj.set_default()
-        prj.data = None
-        prj.add_residential(
-            method='tabula_de',
-            usage='single_family_house',
-            name="ResidentialBuilding",
-            year_of_construction=1858,
-            number_of_floors=2,
-            height_of_floors=3.2,
-            net_leased_area=219)
 
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].outer_walls[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].outer_walls[-1].area),
-            1) == round((1 / (1 / 2.0 - 0.17)), 1)
 
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].ground_floors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].ground_floors[-1].area),
-            1) == round((1 / (1 / 2.9 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].windows[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].windows[-1].area),
-            1) == round((1 / (1 / 2.8 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].doors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].doors[-1].area),
-            1) == round((1 / (1 / 3.0 - 0.17)), 1)
-
-        prj.add_residential(
-            method='tabula_de',
-            usage='single_family_house',
-            name="ResidentialBuilding",
-            year_of_construction=1918,
-            number_of_floors=2,
-            height_of_floors=3.2,
-            net_leased_area=219)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].outer_walls[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].outer_walls[-1].area),
-            1) == round((1 / (1 / 1.7 - 0.17)), 1)
-
-        for floor in prj.buildings[-1].thermal_zones[-1].ground_floors:
-            if "_1_" in floor.construction_type:
-                assert round(1 / (floor.r_conduc * floor.area),
-                             1) == round((1 / (1 / 0.88 - 0.17)), 1)
-            elif "_2_" in floor.construction_type:
-                assert round(1 / (floor.r_conduc * floor.area),
-                             1) == round((1 / (1 / 1.2 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].windows[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].windows[-1].area),
-            1) == round((1 / (1 / 2.8 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].doors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].doors[-1].area),
-            1) == round((1 / (1 / 3.0 - 0.17)), 1)
-
-        prj.add_residential(
-            method='tabula_de',
-            usage='single_family_house',
-            name="ResidentialBuilding",
-            year_of_construction=1947,
-            number_of_floors=2,
-            height_of_floors=3.2,
-            net_leased_area=219)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].outer_walls[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].outer_walls[-1].area),
-            1) == round((1 / (1 / 1.7 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].ground_floors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].ground_floors[-1].area),
-            1) == round((1 / (1 / 0.77 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].windows[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].windows[-1].area),
-            1) == round((1 / (1 / 2.8 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].doors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].doors[-1].area),
-            1) == round((1 / (1 / 3.0 - 0.17)), 1)
-
-        prj.add_residential(
-            method='tabula_de',
-            usage='single_family_house',
-            name="ResidentialBuilding",
-            year_of_construction=1956,
-            number_of_floors=2,
-            height_of_floors=3.2,
-            net_leased_area=219)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].outer_walls[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].outer_walls[-1].area),
-            1) == round((1 / (1 / 1.4 - 0.17)), 1)
-
-        for floor in prj.buildings[-1].thermal_zones[-1].ground_floors:
-            if "_1_" in floor.construction_type:
-                assert round(1 / (floor.r_conduc * floor.area),
-                             1) == round((1 / (1 / 0.78 - 0.17)), 1)
-            elif "_2_" in floor.construction_type:
-                assert round(1 / (floor.r_conduc * floor.area),
-                             1) == round((1 / (1 / 1.0 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].windows[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].windows[-1].area),
-            1) == round((1 / (1 / 2.8 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].doors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].doors[-1].area),
-            1) == round((1 / (1 / 3.0 - 0.17)), 1)
-
-        prj.add_residential(
-            method='tabula_de',
-            usage='single_family_house',
-            name="ResidentialBuilding",
-            year_of_construction=1967,
-            number_of_floors=2,
-            height_of_floors=3.2,
-            net_leased_area=219)
-
-        for wall in prj.buildings[-1].thermal_zones[-1].outer_walls:
-
-            if "_1_" in wall.construction_type:
-                assert round(1 / (wall.r_conduc * wall.area),
-                             1) == round((1 / (1 / 1.2 - 0.17)), 1)
-            """
-            if "_2_" in wall.construction_type:
-                assert round(1 / (wall.r_conduc * wall.area),
-                             1) == round((1 / (1 / 0.8 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].ground_floors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].ground_floors[-1].area),
-            1) == round((1 / (1 / 1.08 - 0.17)), 1)
-            """
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].windows[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].windows[-1].area),
-            1) == round((1 / (1 / 2.8 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].doors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].doors[-1].area),
-            1) == round((1 / (1 / 3.0 - 0.17)), 1)
-
-        prj.add_residential(
-            method='tabula_de',
-            usage='single_family_house',
-            name="ResidentialBuilding",
-            year_of_construction=1977,
-            number_of_floors=2,
-            height_of_floors=3.2,
-            net_leased_area=219)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].outer_walls[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].outer_walls[-1].area),
-            1) == round((1 / (1 / 1.0 - 0.17)), 1)
-
-        for floor in prj.buildings[-1].thermal_zones[-1].ground_floors:
-            if "_1_" in floor.construction_type:
-                assert round(1 / (floor.r_conduc * floor.area),
-                             1) == round((1 / (1 / 0.77 - 0.17)), 1)
-            elif "_2_" in floor.construction_type:
-                assert round(1 / (floor.r_conduc * floor.area),
-                             1) == round((1 / (1 / 1.0 - 0.17)), 1)
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].windows[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].windows[-1].area),
-            1) == round((1 / (1 / 2.8 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].doors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].doors[-1].area),
-            1) == round((1 / (1 / 3.0 - 0.17)), 1)
-
-        prj.add_residential(
-            method='tabula_de',
-            usage='single_family_house',
-            name="ResidentialBuilding",
-            year_of_construction=1982,
-            number_of_floors=2,
-            height_of_floors=3.2,
-            net_leased_area=219)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].outer_walls[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].outer_walls[-1].area),
-            1) == round((1 / (1 / 0.8 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].ground_floors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].ground_floors[-1].area),
-            1) == round((1 / (1 / 0.65 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].windows[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].windows[-1].area),
-            1) == round((1 / (1 / 4.3 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].doors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].doors[-1].area),
-            1) == round((1 / (1 / 3.0 - 0.17)), 1)
-
-        prj.add_residential(
-            method='tabula_de',
-            usage='single_family_house',
-            name="ResidentialBuilding",
-            year_of_construction=1993,
-            number_of_floors=2,
-            height_of_floors=3.2,
-            net_leased_area=219)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].outer_walls[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].outer_walls[-1].area),
-            1) == round((1 / (1 / 0.5 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].ground_floors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].ground_floors[-1].area),
-            1) == round((1 / (1 / 0.51 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].windows[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].windows[-1].area),
-            1) == round((1 / (1 / 3.2 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].doors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].doors[-1].area),
-            1) == round((1 / (1 / 3.0 - 0.17)), 1)
-
-        prj.add_residential(
-            method='tabula_de',
-            usage='single_family_house',
-            name="ResidentialBuilding",
-            year_of_construction=2000,
-            number_of_floors=2,
-            height_of_floors=3.2,
-            net_leased_area=219)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].outer_walls[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].outer_walls[-1].area),
-            1) == round((1 / (1 / 0.3 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].ground_floors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].ground_floors[-1].area),
-            1) == round((1 / (1 / 0.4 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].windows[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].windows[-1].area),
-            1) == round((1 / (1 / 1.9 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].doors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].doors[-1].area),
-            1) == round((1 / (1 / 2.0 - 0.17)), 1)
-
-        prj.add_residential(
-            method='tabula_de',
-            usage='single_family_house',
-            name="ResidentialBuilding",
-            year_of_construction=2008,
-            number_of_floors=2,
-            height_of_floors=3.2,
-            net_leased_area=219)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].outer_walls[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].outer_walls[-1].area),
-            1) == round((1 / (1 / 0.3 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].ground_floors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].ground_floors[-1].area),
-            1) == round((1 / (1 / 0.28 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].windows[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].windows[-1].area),
-            1) == round((1 / (1 / 1.4 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].doors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].doors[-1].area),
-            1) == round((1 / (1 / 2.0 - 0.17)), 1)
-
-        prj.add_residential(
-            method='tabula_de',
-            usage='single_family_house',
-            name="ResidentialBuilding",
-            year_of_construction=2014,
-            number_of_floors=2,
-            height_of_floors=3.2,
-            net_leased_area=219)
-        """
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].outer_walls[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].outer_walls[-1].area),
-            1) == round((1 / (1 / 0.28 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].ground_floors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].ground_floors[-1].area),
-            1) == round((1 / (1 / 0.35 - 0.17)), 1)
-        """
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].windows[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].windows[-1].area),
-            1) == round((1 / (1 / 1.3 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].doors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].doors[-1].area),
-            1) == round((1 / (1 / 1.8 - 0.17)), 1)
-
-        prj.add_residential(
-            method='tabula_de',
-            usage='single_family_house',
-            name="ResidentialBuilding",
-            year_of_construction=2099,
-            number_of_floors=2,
-            height_of_floors=3.2,
-            net_leased_area=219)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].outer_walls[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].outer_walls[-1].area),
-            1) == round((1 / (1 / 0.17 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].ground_floors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].ground_floors[-1].area),
-            1) == round((1 / (1 / 0.17 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].windows[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].windows[-1].area),
-            1) == round((1 / (1 / 1.1 - 0.17)), 1)
-
-        assert round(
-            1 / (prj.buildings[-1].thermal_zones[-1].doors[-1].r_conduc
-                 * prj.buildings[-1].thermal_zones[-1].doors[-1].area),
-            1) == round((1 / (1 / 1.3 - 0.17)), 1)
 
     def test_tabula_uvalue_retrofit_sfh(self):
 
@@ -2031,7 +1668,6 @@ class Test_tabula(object):
             1 / (prj.buildings[-1].thermal_zones[-1].doors[-1].r_conduc
                  * prj.buildings[-1].thermal_zones[-1].doors[-1].area),
             1) == round((1 / (1 / 0.8 - 0.17)), 1)
-'''
 
     def test_tabula_uvalue_standard_th(self):
         prj.set_default()
@@ -3073,3 +2709,4 @@ class Test_tabula(object):
             1 / (prj.buildings[-1].thermal_zones[-1].doors[-1].r_conduc
                  * prj.buildings[-1].thermal_zones[-1].doors[-1].area),
             1) == round((1 / (1 / 0.8 - 0.17)), 1)
+'''
